+++ lib/src/presentation/home_screen/view/home_tab_screen.dart
@@ -7,8 +7,8 @@ import 'package:habitual/src/presentation/home_screen/widgets/home_category_card
 import 'package:habitual/src/presentation/home_screen/widgets/my_interests_card.dart';
 import 'package:habitual/src/routes/app_pages.dart';
 
-import '../../core/core_export.dart';
-import 'widgets/main_card.dart';
+import '../../../core/core_export.dart';
+import '../widgets/main_card.dart';
 
 class HomeTabScreen extends StatefulWidget {
   const HomeTabScreen({super.key});
