+++ lib/src/presentation/splash_screen/view/splash_screen.dart
